refactor: make 'event_wait_info' a non-template base class of promise
This allows us to get rid of the 'waits_on' pointer in 'awaitable' while
still keeping the full promise type erased.

Additionally this reduces the "object access path" to the meta data of
the callgraph.

Then rename the 'handle' representing the waiting coroutine to the
now-available name 'waits_on_me' for increased readability.