performance: reduce memory usage by using SBO for our callgraph
Because almost all coroutines only wait on a single downstream coroutine
or a single I/O event paying for the overhead of dynamic allocation
everytime is undesirable.

So I've added a mini vector implementation that uses the
Small-Buffer Optimization (SBO). This is now used for storing the lists
of 'callees' and 'events' a coroutine is waiting on inside its promise.