feat: add variant_ptr
A space-efficient "variant" of std::variant only for pointers. It uses
the bits left unused due to the alignment of the pointed-to-types.

As a result the number of possible types to store in a variant is
limited by the minimum alignment across all those types.

This implementation uses a union for storage on C++23. When
"P2738: constexpr cast from void*" (C++26) is available it uses a void*
pointer instead. This allows accessing more than the *first* type at
consteval time.

Additionally it plays with "P3125: constexpr pointer tagging" when the
associated experimental Clang builtins are available. This makes
basically everything constexpr.